This time it really seems to work
There are two methods, alphaBetaSearch(Written together yesterday), and alphaBeta (which I have written now and seems working). The check for the turn should be done outside the for loop as in the new method. The check for pruning was wrong, we have to check for apha>=beta, not for the value of the node.